Automation to improve temperature comfort
Based on the average temperature in ground floor and first floor
increase the climate temperature to turn on heating when off and comfort
threshold is not reached in ground floor or first floor. This usually
happens when I turn on the chimney in the ground floor causes the
heating to stop while the first floor is cold.